Update duckyinpython.py
Moving "No Storage" pin to GPIO 10 in junction with moving Setup pin to GPIO 9 with GND as common in junction with a SPDT micro switch to switch between stealth mode and setup mode. Part 2

Also changed the GPIO of Payload 3 and Payload 4 to GPIO 6 & 7 respectively for those options. (GND sits between 4,5 & 6,7. will use this with a custom SP4T switch)

Needs Testing.